base CRUD

+++ src/main/java/com/infotera/teste/repository/PersonRepository.java
@@ -3,11 +3,11 @@ package com.infotera.teste.repository;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.persistence.TypedQuery;
 
 import com.infotera.teste.model.Person;
 
 import java.util.List;
-import java.util.UUID;
 
 @Stateless
 public class PersonRepository {
@@ -22,14 +22,16 @@ public class PersonRepository {
     }
 
     public void delete(Person person) {
-        if (entityManager.contains(person)) {
-            entityManager.remove(person);
-        } else {
-            Person managedPerson = entityManager.find(Person.class, person.getId());
-            if (managedPerson != null) {
-                entityManager.remove(managedPerson);
+    	if(person != null) {
+            if (entityManager.contains(person)) {
+                entityManager.remove(person);
+            } else {
+                Person managedPerson = entityManager.find(Person.class, person.getId());
+                if (managedPerson != null) {
+                    entityManager.remove(managedPerson);
+                }
             }
-        }
+    	}
     }
 
     public void addNewPerson(Person person) {
@@ -47,4 +49,15 @@ public class PersonRepository {
     public void update(List<Person> persons) {
     	persons.forEach(entityManager::merge);
     }
+    
+	public List<Person> searchByName(String name) {
+		String jpql = "from Person where name like :name";
+		
+		TypedQuery<Person> query = entityManager
+				.createQuery(jpql, Person.class);
+		
+		query.setParameter("name", name + "%");
+		
+		return query.getResultList();
+	}
 }
